Dynamic font size (#13) -- resolves #6, #9, #10, #11, #12

+++ src/app/components/edit-round/edit-round.component.ts
@@ -2,23 +2,37 @@ import { CommonModule, DatePipe } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
+import { provideNativeDateAdapter } from '@angular/material/core';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatTableModule } from '@angular/material/table';
 import { Router } from '@angular/router';
-import { APP_ROUTES, NAVIGATION_STATE_KEYS } from '../../models/constants';
+import { TypedTemplateDirective } from '../../directives/typed-template.directive';
+import {
+  APP_ROUTES,
+  DELETE_ROUND,
+  NAVIGATION_STATE_KEYS,
+} from '../../models/constants';
 import { Course } from '../../models/course';
-import { Round } from '../../models/round';
-import { ParPipe } from '../../pipes/par.pipe';
+import { Round, RoundVariety } from '../../models/round';
+import { PipesModule } from '../../pipes/pipes.module';
 import { AppStateService } from '../../services/app-state.service';
 import { CourseService } from '../../services/course.service';
 import { RoundService } from '../../services/round.service';
-import { TotalPuttsPipe } from '../../pipes/total-putts.pipe';
-import { ScoreToParPipe } from '../../pipes/score-to-par.pipe';
-import { TotalRoundScorePipe } from '../../pipes/total-round-score.pipe';
-import { TypedTemplateDirective } from '../../directives/typed-template.directive';
+
+import {
+  MatDatepickerInputEvent,
+  MatDatepickerModule,
+} from '@angular/material/datepicker';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { AreYouSureDialogComponent } from '../are-you-sure-dialog/are-you-sure-dialog.component';
+
+interface ColumnDef {
+  columnDef: string;
+  header: string;
+}
 
 @Component({
   selector: 'app-edit-round',
@@ -30,16 +44,15 @@ import { TypedTemplateDirective } from '../../directives/typed-template.directiv
     MatTableModule,
     MatIconModule,
     MatInputModule,
-    ParPipe,
-    ScoreToParPipe,
+    PipesModule,
     MatSelectModule,
+    MatDatepickerModule,
     DatePipe,
-    TotalRoundScorePipe,
-    TotalPuttsPipe,
     CommonModule,
-    TypedTemplateDirective
+    TypedTemplateDirective,
+    MatDialogModule,
   ],
-  providers: [DatePipe],
+  providers: [DatePipe, provideNativeDateAdapter()],
   templateUrl: './edit-round.component.html',
   styleUrl: './edit-round.component.scss',
 })
@@ -48,10 +61,11 @@ export class EditRoundComponent {
   public coursesToChooseFrom: Course[];
   public currentCourse: Course | null = null;
   public roundIdToEdit: string;
+  public readonly BACK_NINE = RoundVariety.BACK_NINE;
   public readonly HOLE_COL = 'hole';
   public readonly STROKES_COL = 'par';
   public readonly PUTTS_COL = 'putts';
-  public readonly ROUND_TABLE_COLUMNS = [
+  public readonly ROUND_TABLE_COLUMNS: ColumnDef[] = [
     {
       columnDef: this.HOLE_COL,
       header: 'Hole',
@@ -69,15 +83,21 @@ export class EditRoundComponent {
     (def) => def.columnDef
   );
   public SCORE_GRAPHIC_TYPES!: {
-    score: number,
-    scoreToPar: number
-  }
+    score: number;
+    scoreToPar: number;
+  };
+  public HOLE_ROW_TYPES!: {
+    holeIndex: number;
+    column: ColumnDef;
+  };
+  public readonly ROUND_VARIETIES = Object.values(RoundVariety);
 
   constructor(
     private appStateService: AppStateService,
     private courseService: CourseService,
     private roundService: RoundService,
     private router: Router,
+    private dialog: MatDialog,
     datePipe: DatePipe
   ) {
     this.coursesToChooseFrom = this.courseService.getAllCoursesForCurrentUser();
@@ -102,6 +122,7 @@ export class EditRoundComponent {
         putts: new Array(18).fill(undefined),
         courseId: '',
         dateStringISO: new Date().toISOString(), // TODO: make editable
+        roundVariety: RoundVariety.EIGHTEEN,
       };
       this.appStateService.setPageTitle(`Create Round`);
     }
@@ -110,6 +131,11 @@ export class EditRoundComponent {
   public updateCurrentCourse(newCourseId: string): void {
     console.log('selected course', newCourseId);
     this.currentCourse = this.courseService.getCourse(newCourseId);
+  }
+
+  public updateRoundVariety(newRoundVariety: RoundVariety): void {
+    console.log('selected round variety', newRoundVariety);
+    this.editingRound.roundVariety = newRoundVariety;
   }
 
   public strokesPlusOne(index: number) {
@@ -137,21 +163,47 @@ export class EditRoundComponent {
     }
   }
 
+  public addEvent(event: MatDatepickerInputEvent<Date>): void {
+    if (event.value) {
+      this.editingRound.dateStringISO = event.value.toISOString();
+    }
+  }
+
   public get disableSaveButton(): boolean {
     return (
+      !this.editingRound.dateStringISO ||
+      !this.editingRound.roundVariety ||
       !this.editingRound.courseId.length ||
-      this.editingRound.strokes.some((hole) => (hole || 0) <= 0)
+      this.editingRound.strokes.some(
+        (hole, index) =>
+          (hole || 0) <= 0 &&
+          ((index < 9 &&
+            this.editingRound.roundVariety !== RoundVariety.BACK_NINE) ||
+            (index >= 9 &&
+              this.editingRound.roundVariety !== RoundVariety.FRONT_NINE))
+      )
     );
   }
 
   public deleteRound(): void {
-    const updatedCurrentUser = this.appStateService.currentUser;
-    if (updatedCurrentUser) {
-      updatedCurrentUser.roundIds = updatedCurrentUser.roundIds.filter(roundId => roundId !== this.roundIdToEdit);
-    }
-    this.appStateService.currentUser = updatedCurrentUser;
-    this.roundService.deleteRounds([this.roundIdToEdit]);
-    this.router.navigateByUrl(APP_ROUTES.HOME);
+    this.dialog
+      .open(AreYouSureDialogComponent, {
+        data: DELETE_ROUND,
+      })
+      .afterClosed()
+      .subscribe((confirmed) => {
+        if (confirmed) {
+          const updatedCurrentUser = this.appStateService.currentUser;
+          if (updatedCurrentUser) {
+            updatedCurrentUser.roundIds = updatedCurrentUser.roundIds.filter(
+              (roundId) => roundId !== this.roundIdToEdit
+            );
+          }
+          this.appStateService.currentUser = updatedCurrentUser;
+          this.roundService.deleteRounds([this.roundIdToEdit]);
+          this.router.navigateByUrl(APP_ROUTES.HOME);
+        }
+      });
   }
 
   public saveRound(): void {
